fix: move dx12 to support category
as requested by oim1: move dx12 command to support
