+++ src/commands/support/directx12.ts
@@ -5,7 +5,7 @@ import { makeEmbed } from '../../lib/embed';
 export const directx12: CommandDefinition = {
     name: 'dx12',
     description: 'Explaination for the current state of MSFS DirectX12 support.',
-    category: CommandCategory.GENERAL,
+    category: CommandCategory.SUPPORT,
     executor: (msg) => msg.channel.send(makeEmbed({
         title: 'MSFS | DirectX 12',
         description: 'Use of DX11 is still recommended for best performance due to an issue with some glass cockpit technologies in the DX12 version of the simulator. '
